Port to 1.21.2+ (Neo) / 1.21.9+ (Fabric)
Don't ask me why this works on .2 and up for neo but only .9 and up for
fabric. I think it has something to do with the clientside `Input`
class, which appears to have moved in .2 but on fabric it's like it
moved in .9

Also better mixin composability with mixinextras, thanks to unilock!